+++ packages/dash-table/tests/selenium/test_derived_props.py
@@ -135,6 +135,7 @@ def test_tdrp001_select_rows(test):
     assert test.find_element("#derived_virtual_row_ids").get_attribute(
         "innerHTML"
     ) == json.dumps(list(range(3000, 3100)))
+    assert test.get_log_errors() == []
 
 
 def test_tdrp002_select_cell(test):
@@ -192,6 +193,7 @@ def test_tdrp002_select_cell(test):
     assert test.find_element("#derived_virtual_row_ids").get_attribute(
         "innerHTML"
     ) == json.dumps(list(range(3000, 3100)))
+    assert test.get_log_errors() == []
 
 
 def test_tdrp003_select_cells(test):
@@ -326,6 +328,7 @@ def test_tdrp003_select_cells(test):
     assert test.find_element("#derived_virtual_row_ids").get_attribute(
         "innerHTML"
     ) == json.dumps(list(range(3000, 3100)))
+    assert test.get_log_errors() == []
 
 
 def test_tdrp004_navigate_selected_cells(test):
@@ -402,6 +405,8 @@ def test_tdrp004_navigate_selected_cells(test):
             ) == json.dumps(list(range(3000, 3100)))
 
             test.send_keys(Keys.TAB)
+
+    assert test.get_log_errors() == []
 
 
 def test_tdrp005_filtered_and_sorted_row_select(test):
@@ -565,3 +570,4 @@ def test_tdrp005_filtered_and_sorted_row_select(test):
     assert test.find_element("#derived_virtual_row_ids").get_attribute(
         "innerHTML"
     ) == json.dumps(list(range(3000, 3100, 2))[::-1])
+    assert test.get_log_errors() == []