+++ packages/dash-table/tests/selenium/test_markdown_link.py
@@ -63,3 +63,5 @@ def test_tmdl001_click_markdown_link(test, markdown_options, new_tab, cell_selec
     else:
         assert len(test.driver.window_handles) == 1
         assert test.driver.current_url.startswith("https://www.google.com")
+
+    assert test.get_log_errors() == []