+++ packages/dash-table/tests/selenium/test_table_export_xlsx.py
@@ -5,7 +5,7 @@ import dash
 import dash.testing.wait as wait
 
 
-def test_tbex001_table_export(dash_duo):
+def test_tbex001_table_export(test):
     df = pd.read_csv(
         "https://raw.githubusercontent.com/plotly/datasets/master/solar.csv"
     )
@@ -14,13 +14,14 @@ def test_tbex001_table_export(dash_duo):
         id="table",
         columns=[{"name": i, "id": i} for i in df.columns],
         data=df.to_dict("records"),
-        export_format="csv",
+        export_format="xlsx",
     )
-    dash_duo.start_server(app)
-    dash_duo.wait_for_element(".export", timeout=1).click()
+    test.start_server(app)
+    test.wait_for_element(".export").click()
 
-    download = os.path.sep.join((dash_duo.download_path, "Data.csv"))
+    download = os.path.sep.join((test.download_path, "Data.xlsx"))
     wait.until(lambda: os.path.exists(download), timeout=2)
 
-    df_bis = pd.read_csv(download)
+    df_bis = pd.read_excel(download)
     assert df_bis.equals(df)
+    assert test.get_log_errors() == []